+jwt-authorization

+++ resources/item.py
@@ -1,22 +1,26 @@
-import uuid
-from flask import request
+
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
 from schemas import ItemSchema , ItemUpdateSchema 
 from models import ItemModel
+from flask_jwt_extended import jwt_required , get_jwt   
 from db import db
 from sqlalchemy.exc import SQLAlchemyError
 blp = Blueprint("Items", __name__, description="Operations on items")
 
 
-@blp.route("/item/<string:item_id>")
+@blp.route("/item/<int:item_id>")
 class Item(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):
         item = ItemModel.query.get_or_404(item_id)
         return item 
-    
+    @jwt_required()
     def delete(self, item_id):
+        jwt= get_jwt()
+        if not jwt.get("is_admin"):
+            abort(401, message="Admin privilege required.")
         item = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit() 
@@ -25,7 +29,7 @@ class Item(MethodView):
     
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchema)
-
+    @jwt_required()
     def put(self,item_data ,item_id):
         item = ItemModel.query.get(item_id)
         if item :
@@ -38,11 +42,14 @@ class Item(MethodView):
         return item
 
 @blp.route("/item")
+
 class ItemList(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self):
         return ItemModel.query.all()
 
+    @jwt_required(fresh=True)
     @blp.arguments(ItemSchema)
     @blp.response(201, ItemSchema)
     def post(self , item_data):
